Minor refactoring for styling
