+++ Carbon/config/Api.js
@@ -1,5 +1,5 @@
 // Testing Server
-export const API_URL = "http://localhost:3000/api/"
+//export const API_URL = "http://localhost:3000/api/"
 
 // Deployed Server
-// export const API_URL = "http://Carbonserver-env.eba-pmpdtmpe.us-east-1.elasticbeanstalk.com/api/"
+export const API_URL = "http://Carbonserver-env.eba-pmpdtmpe.us-east-1.elasticbeanstalk.com/api/"
