+++ Carbon/navigation/screens/Progress/GoalSetter.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { View, Text, StyleSheet, TouchableOpacity } from 'react-native';
 import Slider from '@react-native-community/slider';
-import { Colors } from '../../../colors/Colors';
+import { Colors } from '../../../styling/Colors';
 import { ScreenNames } from '../Main/ScreenNames';
 import { API_URL } from '../../../config/Api';
 import { getToken } from '../../../util/LoginManager';
